requirements.txt
# Web framework
jinja2>=3.1.0
aiofiles>=23.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Validation / Serialisation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Task queue
redis>=5.0.0
rq>=1.16.0

# PDF processing
pdf2image>=1.17.0
pdfplumber>=0.9.0
camelot-py[base]>=0.11.0

# OCR
pytesseract>=0.3.10
google-cloud-documentai>=2.24.0

# Image processing
Pillow>=10.2.0
opencv-python-headless>=4.9.0

# Date parsing
python-dateutil>=2.8.2

# File upload
python-multipart>=0.0.6

# Observability
structlog>=24.1.0
sentry-sdk[fastapi]>=1.40.0
prometheus-client>=0.20.0

# HTTP client (webhooks)
httpx>=0.27.0

# Numeric / stats
numpy>=1.26.0
scipy>=1.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Linting
ruff>=0.2.0